Make usage statistics calculation a subprocess so it doesn't slow site
